Add compression of arbitrary length strings.
Compression is made by blocks. The compression loop repeats until the difference between
the b_inf and b_sup is comparable to std::f64::EPSILON.
As such I obtained to compress on average blocks of 8 to 13 bytes into a 8 bytes floating number.
Compression ratio obtained is generally around 0.70.

Problem to solve, the uncompress func still needs to know how much bytes it needs to
uncompress for each block. There must be a way to pass this info within the compressed data.